analytics_service/backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.3
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.2
pydantic-settings==2.4.0
httpx==0.27.2
python-dotenv==1.0.1
orjson==3.10.7
tenacity==9.0.0

